since we will use always the header in the layout i introduced it in
