+++ src/screens/layout/Layout.tsx
@@ -1,6 +1,7 @@
 import { Outlet, useLocation } from "react-router-dom";
 import { ReactComponent as HuginnIcon } from "../../assets/huginn_logo_icon.svg";
 import { useState } from "react";
+import { Header } from "../../components/Header";
 
 export const Layout = () => {
   const { pathname } = useLocation();
@@ -13,10 +14,21 @@ export const Layout = () => {
     setChatIsOpen(true);
   }
 
+  const handleOpenFullWindowChat = () => {
+    const chatUrl = window.location.origin + "/full-screen";
+    window.open(chatUrl, "_blank");
+  };
+
   return (
     <div className={`app-container ${path}`}>
       {shouldShowOutlet && (
         <div className={`chatbot-container ${path}`}>
+          <Header
+            handleOpenChat={() =>
+              setChatIsOpen((prevChatIsOpen: boolean) => !prevChatIsOpen)
+            }
+            handleOpenFullWindowChat={handleOpenFullWindowChat}
+          />
           <Outlet context={{ chatIsOpen, setChatIsOpen }} />
         </div>
       )}
